requirements.txt
flask==2.2.3
werkzeug==2.2.3
langchain==0.0.267
langchain_core==0.1.5
openai==0.27.8
python-dotenv==1.0.0
flask-socketio==5.3.4
eventlet==0.33.3
gTTS==2.3.2
SpeechRecognition==3.10.0
pygame==2.5.2
PyAudio==0.2.13 